+++ src/lib.rs
@@ -88,6 +88,7 @@
 //!         .with_validators(4)
 //!         .with_service(TimestampingService)
 //!         .create();
+//!
 //!     // Create few transactions.
 //!     let keypair = gen_keypair();
 //!     let tx1 = TxTimestamp::new(&keypair.0, "Down To Earth", &keypair.1);
@@ -97,16 +98,23 @@
 //!     testkit.create_block_with_transactions(txvec![
 //!         tx1.clone(), tx2.clone(), tx3.clone()
 //!     ]);
+//!
+//!     // Add a single transaction.
+//!     let tx4 = TxTimestamp::new(&keypair.0, "Barking up the wrong tree", &keypair.1);
+//!     testkit.create_block_with_transaction(tx4.clone());
+//!
 //!     // Check results with schema.
 //!     let snapshot = testkit.snapshot();
 //!     let schema = Schema::new(&snapshot);
 //!     assert!(schema.transactions().contains(&tx1.hash()));
 //!     assert!(schema.transactions().contains(&tx2.hash()));
 //!     assert!(schema.transactions().contains(&tx3.hash()));
+//!     assert!(schema.transactions().contains(&tx4.hash()));
+//!
 //!     // Check results with api.
 //!     let api = testkit.api();
 //!     let blocks: Vec<Block> = api.get(ApiKind::Explorer, "v1/blocks?count=10");
-//!     assert_eq!(blocks.len(), 2);
+//!     assert_eq!(blocks.len(), 3);
 //!     api.get::<serde_json::Value>(
 //!         ApiKind::System,
 //!         &format!("v1/transactions/{}", tx1.hash().to_string()),
@@ -146,16 +154,19 @@ use exonum::crypto;
 use exonum::helpers::{Height, Round, ValidatorId};
 use exonum::messages::{Message, Precommit, Propose};
 use exonum::node::{ApiSender, ExternalMessage, State as NodeState, TransactionSend, TxPool};
-use exonum::storage::{Database, MemoryDB, Snapshot};
+use exonum::storage::{MemoryDB, Snapshot};
 
 #[macro_use]
 mod macros;
+mod checkpoint_db;
 pub mod compare;
 mod greedy_fold;
 
 #[doc(hidden)]
 pub use greedy_fold::GreedilyFoldable;
 pub use compare::ComparableSnapshot;
+
+use checkpoint_db::{CheckpointDb, CheckpointDbHandler};
 
 /// Emulated test network.
 #[derive(Debug)]
@@ -328,12 +339,12 @@ impl TestNode {
         self.validator_id
     }
 
-    /// Change node role.
+    /// Changes node role.
     pub fn change_role(&mut self, role: Option<ValidatorId>) {
         self.validator_id = role;
     }
 
-    /// Returns the service keypar.
+    /// Returns the service keypair.
     pub fn service_keypair(&self) -> (&crypto::PublicKey, &crypto::SecretKey) {
         (&self.service_public_key, &self.service_secret_key)
     }
@@ -445,9 +456,7 @@ impl TestKitBuilder {
     /// Creates the testkit.
     pub fn create(self) -> TestKit {
         crypto::init();
-        let db = MemoryDB::new();
         TestKit::assemble(
-            Box::new(db),
             self.services,
             TestNetwork {
                 us: self.us,
@@ -461,6 +470,7 @@ impl TestKitBuilder {
 /// (with no real network setup).
 pub struct TestKit {
     blockchain: Blockchain,
+    db_handler: CheckpointDbHandler<MemoryDB>,
     events_stream: Spawn<Box<Stream<Item = (), Error = ()>>>,
     network: TestNetwork,
     api_sender: ApiSender,
@@ -495,12 +505,15 @@ impl TestKit {
         (testkit, public_key, secret_key)
     }
 
-    fn assemble(db: Box<Database>, services: Vec<Box<Service>>, network: TestNetwork) -> Self {
+    fn assemble(services: Vec<Box<Service>>, network: TestNetwork) -> Self {
         let api_channel = mpsc::channel(1_000);
         let api_sender = ApiSender::new(api_channel.0.clone());
 
+        let db = CheckpointDb::new(MemoryDB::new());
+        let db_handler = db.handler();
+
         let mut blockchain = Blockchain::new(
-            db,
+            Box::new(db),
             services,
             *network.us().service_keypair().0,
             network.us().service_keypair().1.clone(),
@@ -535,6 +548,7 @@ impl TestKit {
 
         TestKit {
             blockchain,
+            db_handler,
             api_sender,
             events_stream,
             network,
@@ -574,64 +588,104 @@ impl TestKit {
         &mut self.blockchain
     }
 
+    /// Rolls the blockchain back for a certain number of blocks.
+    ///
+    /// # Examples
+    ///
+    /// Rollbacks are useful in testing alternative scenarios (e.g., transactions executed
+    /// in different order and/or in different blocks) that require an expensive setup:
+    ///
+    /// ```
+    /// # #[macro_use] extern crate exonum;
+    /// # #[macro_use] extern crate exonum_testkit;
+    /// # use exonum::blockchain::{Service, Transaction};
+    /// # use exonum::messages::RawTransaction;
+    /// # use exonum::encoding;
+    /// # use exonum_testkit::{TestKit, TestKitBuilder};
+    /// #
+    /// # type FromRawResult = Result<Box<Transaction>, encoding::Error>;
+    /// # pub struct MyService;
+    /// # impl Service for MyService {
+    /// #    fn service_name(&self) -> &'static str {
+    /// #        "documentation"
+    /// #    }
+    /// #    fn service_id(&self) -> u16 {
+    /// #        0
+    /// #    }
+    /// #    fn tx_from_raw(&self, _raw: RawTransaction) -> FromRawResult {
+    /// #        unimplemented!();
+    /// #    }
+    /// # }
+    /// #
+    /// # message! {
+    /// #     struct MyTransaction {
+    /// #         const TYPE = 0;
+    /// #         const ID = 0;
+    /// #         const SIZE = 40;
+    /// #         field from: &exonum::crypto::PublicKey [0 => 32]
+    /// #         field msg: &str [32 => 40]
+    /// #     }
+    /// # }
+    /// # impl Transaction for MyTransaction {
+    /// #     fn verify(&self) -> bool { true }
+    /// #     fn execute(&self, _: &mut exonum::storage::Fork) {}
+    /// # }
+    /// #
+    /// # fn expensive_setup(_: &mut TestKit) {}
+    /// # fn assert_something_about(_: &TestKit) {}
+    /// #
+    /// # fn main() {
+    /// let mut testkit = TestKitBuilder::validator()
+    ///     .with_service(MyService)
+    ///     .create();
+    /// expensive_setup(&mut testkit);
+    /// let (pubkey, key) = exonum::crypto::gen_keypair();
+    /// let tx_a = MyTransaction::new(&pubkey, "foo", &key);
+    /// let tx_b = MyTransaction::new(&pubkey, "bar", &key);
+    /// testkit.create_block_with_transactions(txvec![tx_a.clone(), tx_b.clone()]);
+    /// assert_something_about(&testkit);
+    /// testkit.rollback(1);
+    /// testkit.create_block_with_transactions(txvec![tx_a.clone()]);
+    /// testkit.create_block_with_transactions(txvec![tx_b.clone()]);
+    /// assert_something_about(&testkit);
+    /// testkit.rollback(2);
+    /// # }
+    /// ```
+    pub fn rollback(&mut self, blocks: usize) {
+        assert!(
+            (blocks as u64) <= self.height().0,
+            "Cannot rollback past genesis block"
+        );
+        self.db_handler.rollback(blocks);
+    }
+
     /// Executes a list of transactions given the current state of the blockchain, but does not
     /// commit execution results to the blockchain. The execution result is the same
     /// as if transactions were included into a new block; for example,
     /// transactions included into one of previous blocks do not lead to any state changes.
-    ///
-    /// # Panics
-    ///
-    /// - Panics if there are duplicate transactions.
-    pub fn probe_all(&self, transactions: Vec<Box<Transaction>>) -> Box<Snapshot> {
-        let validator_id = self.network().us().validator_id().expect(
-            "Tested node is not a validator",
-        );
-        let height = self.height().next();
+    pub fn probe_all<I>(&mut self, transactions: I) -> Box<Snapshot>
+    where
+        I: IntoIterator<Item = Box<Transaction>>,
+    {
+        // Filter out already committed transactions; otherwise,
+        // `create_block_with_transactions()` will panic.
+        let schema = CoreSchema::new(self.snapshot());
+        let uncommitted_txs = transactions.into_iter().filter(|tx| {
+            !schema.transactions().contains(&tx.hash())
+        });
 
-        let (transaction_map, hashes) = {
-            let mut transaction_map = BTreeMap::new();
-            let mut hashes = Vec::with_capacity(transactions.len());
-
-            let core_schema = CoreSchema::new(self.snapshot());
-            let committed_txs = core_schema.transactions();
-
-            for tx in transactions {
-                let hash = tx.hash();
-                if committed_txs.contains(&hash) {
-                    continue;
-                }
-
-                hashes.push(hash);
-                transaction_map.insert(hash, tx);
-            }
-
-            assert_eq!(
-                hashes.len(),
-                transaction_map.len(),
-                "Duplicate transactions in probe"
-            );
-
-            (transaction_map, hashes)
-        };
-
-        let (_, patch) = self.blockchain.create_patch(
-            validator_id,
-            height,
-            &hashes,
-            &transaction_map,
-        );
-
-        let mut fork = self.blockchain.fork();
-        fork.merge(patch);
-        Box::new(fork)
+        self.create_block_with_transactions(uncommitted_txs);
+        let snapshot = self.snapshot();
+        self.rollback(1);
+        snapshot
     }
 
     /// Executes a transaction given the current state of the blockchain but does not
     /// commit execution results to the blockchain. The execution result is the same
     /// as if a transaction was included into a new block; for example,
     /// a transaction included into one of previous blocks does not lead to any state changes.
-    pub fn probe<T: Transaction>(&self, transaction: T) -> Box<Snapshot> {
-        self.probe_all(vec![Box::new(transaction)])
+    pub fn probe<T: Transaction>(&mut self, transaction: T) -> Box<Snapshot> {
+        self.probe_all(vec![Box::new(transaction) as Box<Transaction>])
     }
 
     fn do_create_block(&mut self, tx_hashes: &[crypto::Hash]) {
@@ -710,8 +764,8 @@ impl TestKit {
         }
     }
 
-    /// Creates block with the given transactions.
-    /// Transactions that are in mempool will be ignored.
+    /// Creates a block with the given transactions.
+    /// Transactions that are in the mempool will be ignored.
     ///
     /// # Panics
     ///
@@ -742,6 +796,16 @@ impl TestKit {
                 .collect()
         };
         self.create_block_with_tx_hashes(&tx_hashes);
+    }
+
+    /// Creates a block with the given transaction.
+    /// Transactions that are in the mempool will be ignored.
+    ///
+    /// # Panics
+    ///
+    /// - Panics if given transaction has been already committed to the blockchain.
+    pub fn create_block_with_transaction<T: Transaction>(&mut self, tx: T) {
+        self.create_block_with_transactions(txvec![tx]);
     }
 
     /// Creates block with the specified transactions. The transactions must be previously
@@ -1047,11 +1111,18 @@ impl TestNetworkConfiguration {
     }
 }
 
-#[doc(hidden)]
+/// Kind of public or private REST API of an Exonum node.
+///
+/// `ApiKind` allows to use `get*` and `post*` methods of [`TestKitApi`] more safely.
+///
+/// [`TestKitApi`]: struct.TestKitApi.html
 #[derive(Debug)]
 pub enum ApiKind {
+    /// `api/system` endpoints of the built-in Exonum REST API.
     System,
+    /// `api/explorer` endpoints of the built-in Exonum REST API.
     Explorer,
+    /// Endpoints corresponding to a service with the specified string identifier.
     Service(&'static str),
 }
 
@@ -1080,7 +1151,7 @@ impl fmt::Debug for TestKitApi {
 }
 
 impl TestKitApi {
-    /// Creates a new instance of Api.
+    /// Creates a new instance of API.
     fn new(testkit: &TestKit) -> Self {
         use std::sync::Arc;
         use exonum::api::{public, Api};
@@ -1176,6 +1247,11 @@ impl TestKitApi {
     }
 
     /// Gets information from a public endpoint of the node.
+    ///
+    /// # Panics
+    ///
+    /// - Panics if an error occurs during request processing (e.g., the requested endpoint is
+    ///  unknown), or if the response has a non-20x response status.
     pub fn get<D>(&self, kind: ApiKind, endpoint: &str) -> D
     where
         for<'de> D: Deserialize<'de>,
@@ -1188,6 +1264,11 @@ impl TestKitApi {
     }
 
     /// Gets information from a private endpoint of the node.
+    ///
+    /// # Panics
+    ///
+    /// - Panics if an error occurs during request processing (e.g., the requested endpoint is
+    ///  unknown), or if the response has a non-20x response status.
     pub fn get_private<D>(&self, kind: ApiKind, endpoint: &str) -> D
     where
         for<'de> D: Deserialize<'de>,
@@ -1200,6 +1281,10 @@ impl TestKitApi {
     }
 
     /// Gets an error from a public endpoint of the node.
+    ///
+    /// # Panics
+    ///
+    /// - Panics if the response has a non-40x response status.
     pub fn get_err<D>(&self, kind: ApiKind, endpoint: &str) -> D
     where
         for<'de> D: Deserialize<'de>,
@@ -1236,6 +1321,11 @@ impl TestKitApi {
     /// of synchronous transaction processing, which includes running the API shim
     /// and `Transaction.verify()`. `Transaction.execute()` is not run until the transaction
     /// gets to a block via one of `create_block*()` methods.
+    ///
+    /// # Panics
+    ///
+    /// - Panics if an error occurs during request processing (e.g., the requested endpoint is
+    ///  unknown).
     pub fn post<T, D>(&self, kind: ApiKind, endpoint: &str, transaction: &T) -> D
     where
         T: Serialize,
@@ -1252,6 +1342,11 @@ impl TestKitApi {
     /// of synchronous transaction processing, which includes running the API shim
     /// and `Transaction.verify()`. `Transaction.execute()` is not run until the transaction
     /// gets to a block via one of `create_block*()` methods.
+    ///
+    /// # Panics
+    ///
+    /// - Panics if an error occurs during request processing (e.g., the requested endpoint is
+    ///  unknown).
     pub fn post_private<T, D>(&self, kind: ApiKind, endpoint: &str, transaction: &T) -> D
     where
         T: Serialize,